refactor(issue69): move daily_stock_tier_batch under src/pipeline
- Keep src/daily_stock_tier_batch.py as backward-compatible wrapper\n- Move implementation to src/pipeline/daily_stock_tier_batch.py\n- Update pipeline orchestrator import + entrypoint compat guard\n- Fix .gitignore to keep __pycache__ ignored under src/data
