+++ todos/2026_02_16-issue69-src-package-restructure-breakdown.md
@@ -224,6 +224,14 @@
 - [x] 테스트 통과:
   - `python -m unittest tests.test_issue60_import_side_effects tests.test_issue61_import_style_standardization tests.test_issue68_wfo_import_side_effects tests.test_issue69_entrypoint_compat tests.test_pipeline_batch tests.test_ticker_universe_batch tests.test_ohlcv_batch -v`
 
+### 6-5. PR-4: `daily_stock_tier_batch` 구현을 `src/pipeline`로 이동
+- [x] `src/pipeline/daily_stock_tier_batch.py`: 기존 `src/daily_stock_tier_batch.py` 구현 이동(상위 모듈 상대 import로 정리)
+- [x] `src/daily_stock_tier_batch.py`: backward-compatible wrapper로 재생성(주요 심볼 re-export)
+- [x] `src/pipeline/batch.py`: 내부 import를 wrapper 대신 `.daily_stock_tier_batch` 구현으로 전환
+- [x] `tests/test_issue69_entrypoint_compat.py`: `src.daily_stock_tier_batch` import + 심볼(`run_daily_stock_tier_batch`) 가드 추가
+- [x] 테스트 통과:
+  - `python -m unittest tests.test_issue60_import_side_effects tests.test_issue61_import_style_standardization tests.test_issue68_wfo_import_side_effects tests.test_issue69_entrypoint_compat tests.test_pipeline_batch tests.test_daily_stock_tier_batch -v`
+
 ---
 
 ## 7. 문제 해결에 참고
